DEP-3819: Created DAv3 Child Benefit page
