+++ conf/app.routes
@@ -40,9 +40,9 @@ GET         /chat/self-assessment                                               
 GET         /chat/online-services-helpdesk                                              controllers.CuiController.CuiController.onlineServicesHelpdesk
 GET         /chat/employer-enquiries                                                    controllers.CuiController.CuiController.employerEnquiries
 GET         /chat/construction-industry-scheme                                          controllers.CuiController.CuiController.constructionIndustryScheme
-GET         /chat/child-benefit                                                         controllers.CuiController.CuiController.childBenefit
 
 GET         /chat/ask-hmrc-online                                                       controllers.CiapiController.CiapiController.askHmrcOnline
 GET         /chat/customs-and-international-trade                                       controllers.CiapiController.CiapiController.customsInternationalTrade
 GET			/chat/vat-online															controllers.CiapiController.CiapiController.vatOnline
 GET         /chat/corporation-tax                                                       controllers.CiapiController.CiapiController.corporationTax
+GET         /chat/child-benefit                                                         controllers.CiapiController.CiapiController.childBenefit
