+++ test/controllers/CiapiController/CiapiControllerSpec.scala
@@ -103,5 +103,29 @@ class CiapiControllerSpec
         status(result) mustBe NOT_FOUND
       }
     }
+
+    "render child benefit CUI page if showCHBCUI is true" in {
+      val application = builder.configure("features.showCHBCUI" -> "true").build()
+
+      running(application) {
+        val request = FakeRequest(GET, ciapiRoutes.CiapiController.childBenefit.url)
+        val result = route(application, request).get
+        val doc = asDocument(contentAsString(result))
+        status(result) mustBe OK
+        doc.select("h1").text() mustBe "Child Benefit: chat"
+      }
+    }
+
+    "render child benefit DAv2 CUI page if showCHBCUI is false" in {
+      val application = builder.configure("features.showCHBCUI" -> "false").build()
+
+      running(application) {
+        val request = FakeRequest(GET, ciapiRoutes.CiapiController.childBenefit.url)
+        val result = route(application, request).get
+        val doc = asDocument(contentAsString(result))
+        status(result) mustBe OK
+        doc.select("h1").text() mustBe "Child Benefit: chat"
+      }
+    }
   }
 }
