+++ test/controllers/CuiController/CuiControllerSpec.scala
@@ -69,28 +69,5 @@ class CuiControllerSpec
       status(result) mustBe OK
       doc.select("h1").text() mustBe "Construction Industry Scheme: chat"
     }
-
-    "render child benefit CUI page if showCHBCUI is true" in {
-      val application = builder.configure("features.showCHBCUI" -> "true").build()
-
-      running(application) {
-        val request = FakeRequest(GET, cuiRoutes.CuiController.childBenefit.url)
-        val result = route(application, request).get
-        val doc = asDocument(contentAsString(result))
-        status(result) mustBe OK
-        doc.select("h1").text() mustBe "Child Benefit: chat"
-      }
-    }
-
-    "render technical support with HMRC online services page : if showCHBCUI is false" in {
-      val application = builder.configure("features.showCHBCUI" -> "false").build()
-
-      running(application) {
-        val request = FakeRequest(GET, cuiRoutes.CuiController.childBenefit.url)
-        val result = route(application, request).get
-        status(result) mustBe NOT_FOUND
-      }
-    }
-
   }
 }
